dbt/pyproject.toml
[project]
name = "pen-dataprodukt"
version = "0.1.0"
description = "Lage dataprodukt med dbt i PEN, oracledatabasen for pensjon og uføre"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # se https://github.com/navikt/dvh-images/blob/main/airflow/dbt/pyproject.toml
    "dbt-oracle>=1.9.0",
    "oracledb>=2.3.0",
    "requests>=2.32.3",
    "db-dtypes>=1.2.0",
    "dbt-core>=1.9.1",
]


[tool.uv]
package = false

[dependency-groups]
dev = [
    "black>=24.10.0",
    "google-cloud-secret-manager>=2.24.0",
    "sqlfluff>=3.3.0",
]
